RAIR-4656 - Update import contract functionality

* RAIR-4561 - Update import contract functionality


Approved-by: Suresh Arora

+++ rairnode/bin/integrations/ethers/importContractData.js
@@ -16,7 +16,7 @@ const insertToken = async (token, contractId) => {
     log.error(`Error importing token #${token.tokenId}: ${token.metadataError}`);
     return false;
   }
-  if (!token.rawMetadata && token.tokenUri) {
+  if (!token?.raw?.metadata && token.tokenUri) {
     try {
       log.info(`${token.token_id} has no metadata in Moralis, will try to fetch metadata from ${token.tokenUri}`);
       metadata = await (await fetch(token.tokenUri.gateway)).json();
@@ -27,24 +27,18 @@ const insertToken = async (token, contractId) => {
     }
   } else {
     // Use the image loaded in the metadata
-    metadata = token?.rawMetadata;
+    metadata = token.raw.metadata;
+    // If it doesn't exist, use the original URL from the metadata
     if (!metadata.image) {
-      if (token?.metadata?.at) {
-        metadata.image = token?.media?.at(0)?.gateway;
-      } else {
-        metadata.image = token?.media?.gateway;
-      }
-    }
-    // Use the image from Alchemy's gateway
-    // We do it regardless of the value in rawMetadata because some ipfs gateways
-    // get disabled over time, Alchemy's gateway should be more reliable
-    if (!metadata.image) {
-      // Use the raw IPFS link and use the default gateway
-      metadata.image = token?.media?.raw;
+      metadata.image = token?.image?.originalUrl;
     }
     if (metadata.image) {
       // If the ipfs prefix still exists, replace it
       metadata.image = metadata?.image?.replace('ipfs://', 'https://ipfs.io/ipfs/');
+    }
+    // Load image thumbnail
+    if (token?.image?.thumbnailUrl) {
+      metadata.image_thumbnail = token?.image?.thumbnailUrl.replace('ipfs://', 'https://ipfs.io/ipfs/');
     }
   }
   if (metadata && metadata.image && metadata.name && token.owner) {
@@ -68,7 +62,11 @@ const insertToken = async (token, contractId) => {
       }));
     }
     try {
-      await (new MintedToken({
+      // Check if token already exists in database
+      await MintedToken.findOneAndUpdate({
+        contract: contractId,
+        uniqueIndexInContract: token.tokenId,
+      }, {
         ownerAddress: token.owner.toLowerCase(),
         metadataURI: token.token_uri,
         metadata,
@@ -79,9 +77,11 @@ const insertToken = async (token, contractId) => {
         offer: 0,
         offerPool: 0,
         product: 0,
-      })).save();
+      }, {
+        upsert: true
+      });
     } catch (error) {
-      log.error(`Error inserting token ${token.token_id}, ${error.name}`);
+      log.error(`Error upserting token ${token.token_id}, ${error.name}`);
     }
   }
   return true;
@@ -89,7 +89,7 @@ const insertToken = async (token, contractId) => {
 
 module.exports = {
   importContractData: async (networkId, contractAddress, limit, contractCreator, importerUser) => {
-    let contract;
+    let contract, product, offer, offerPool;
 
     // Optional Config object, but defaults to demo api-key and eth-mainnet.
     const settings = {
@@ -101,15 +101,13 @@ module.exports = {
     }
     const alchemySDK = new Alchemy(settings);
 
+    let update = false;
+
     contract = await Contract.findOne({
       contractAddress,
       blockchain: networkId,
       external: true,
     });
-
-    if (contract) {
-      return { success: false, result: undefined, message: 'NFTs already imported' };
-    }
 
     const contractMetadata = await alchemySDK.nft.getContractMetadata(contractAddress);
 
@@ -120,49 +118,53 @@ module.exports = {
       return { success: false, result: undefined, message: `Only ERC721 is supported, tried to process a ${contractMetadata.tokenType} contract` };
     }
 
-    contract = await (new Contract({
-      user: contractCreator,
-      title: contractMetadata.name,
-      contractAddress,
-      blockchain: networkId,
-      importedBy: importerUser,
-      diamond: false,
-      external: true,
-    }));
-
-    const product = await (new Product({
-      name: contractMetadata.name,
-      collectionIndexInContract: 0,
-      contract: contract._id,
-      copies: contractMetadata.totalSupply,
-      soldCopies: contractMetadata.totalSupply,
-      sold: true,
-      firstTokenIndex: 0,
-      transactionHash: 'UNKNOWN - External Import',
-    }));
-
-    const offer = await new Offer({
-      offerIndex: 0,
-      contract: contract._id,
-      product: 0,
-      offerPool: 0,
-      copies: contractMetadata.totalSupply,
-      soldCopies: contractMetadata.totalSupply - 1,
-      sold: true,
-      price: '0',
-      range: [0, contractMetadata.totalSupply],
-      offerName: contractMetadata.name,
-      diamondRangeIndex: 0,
-      transactionHash: 'UNKNOWN - External Import',
-    });
-
-    const offerPool = await (new OfferPool({
-      marketplaceCatalogIndex: 0,
-      contract: contract._id,
-      product: 0,
-      rangeNumber: 0,
-      transactionHash: 'UNKNOWN - External Import',
-    }));
+    if (!contract) {
+      contract = new Contract({
+        user: contractCreator,
+        title: contractMetadata.name,
+        contractAddress,
+        blockchain: networkId,
+        importedBy: importerUser,
+        diamond: false,
+        external: true,
+      });
+      product = new Product({
+        name: contractMetadata.name,
+        collectionIndexInContract: 0,
+        contract: contract._id,
+        copies: contractMetadata.totalSupply,
+        soldCopies: contractMetadata.totalSupply,
+        sold: true,
+        firstTokenIndex: 0,
+        transactionHash: 'UNKNOWN - External Import',
+      });
+      offer = new Offer({
+        offerIndex: 0,
+        contract: contract._id,
+        product: 0,
+        offerPool: 0,
+        copies: contractMetadata.totalSupply,
+        soldCopies: contractMetadata.totalSupply - 1,
+        sold: true,
+        price: '0',
+        range: [0, contractMetadata.totalSupply],
+        offerName: contractMetadata.name,
+        diamondRangeIndex: 0,
+        transactionHash: 'UNKNOWN - External Import',
+      });
+      offerPool = new OfferPool({
+        marketplaceCatalogIndex: 0,
+        contract: contract._id,
+        product: 0,
+        rangeNumber: 0,
+        transactionHash: 'UNKNOWN - External Import',
+      });
+    } else {
+      update = true;
+      product = await Product.findOne({ contract: contract._id });
+      offer = await Offer.findOne({ contract: contract._id });
+      offerPool = await OfferPool.findOne({ contract: contract._id });
+    }
 
     // Can't be used, it doesn't say which NFT they own
     // console.log(await alchemySDK.nft.getOwnersForContract(contractAddress));
@@ -200,11 +202,10 @@ module.exports = {
           contract,
           numberOfTokensAdded,
         },
-        message: '',
       };
     } catch (err) {
       log.error(err);
-      if (contract) {
+      if (contract && !update) {
         MintedToken.deleteMany({ contract: contract._id });
         Offer.deleteMany({ contract: contract._id });
         Product.deleteMany({ contract: contract._id });
